BACKEND/requirements.txt
fastapi[standard]
uvicorn
sqlmodel
psycopg2-binary
pydantic-settings